fixed todos exposed id

+++ src/main/java/com/microservice/user/entity/JobEntity.java
@@ -2,13 +2,12 @@ package com.microservice.user.entity;
 
 import lombok.Data;
 
-//todo: Change int -> Long for jobProviderId and jobReceiverId
 @Data
 public class JobEntity {
 
     private Long jobId;
     private String jobTitle;
     private String jobDescription;
-    private int jobProviderId;
-    private int jobReceiverId;
+    private Long jobProviderId;
+    private Long jobReceiverId;
 }
